requirements.txt
datasets==2.16.1
einops==0.7.0
faiss_gpu==1.7.2
FlagEmbedding==1.2.5
FlagEmbedding==1.2.5
flash_attn==2.5.6
gradio==4.21.0
langchain==0.1.12
langchain_community==0.0.28
langchain_openai==0.0.8
mmengine==0.10.3
mmengine_lite==0.10.3
numpy==1.26.4
pdfplumber==0.10.3
peft==0.8.1
python_docx==1.1.0
sentencepiece==0.1.99
tokenizers==0.15.2
torch==2.1.2
tqdm==4.66.1
transformers==4.38.2
xtuner==0.1.14
